+++ payment-service/src/main/java/com/kyro/payment/dto/PaymentDetailDTO.java
@@ -4,7 +4,7 @@ public record PaymentDetailDTO(
     Long paymentId,
     String paymentMethod,
     String status,
-    int amount,
+    long amount,
     String vnp_TxnRef,
     String vnp_TransactionNo,
     String vnp_ResponseCode) {}
